Simplify dependencies - move dependency in form to form itsefl - do not pass dependencies as an option argument
